+++ imod/tests/conftest.py
@@ -88,4 +88,4 @@ from .fixtures.mf6_welltest_fixture import (
     well_test_data_transient,
 )
 from .fixtures.msw_fixture import fixed_format_parser, simple_2d_grid_with_subunits
-from .fixtures.msw_model_fixture import coupled_mf6_model, msw_model
+from .fixtures.msw_model_fixture import coupled_mf6_model, coupled_mf6wel, msw_model