+++ imod/tests/test_msw/test_initial_conditions.py
@@ -15,14 +15,15 @@ from imod.msw import (
 )
 from imod.typing.grid import is_empty
 
+DUMMY_ARGS = None, None, None, None
+
 
 def test_initial_conditions_equilibrium():
     ic = InitialConditionsEquilibrium()
-    dummy = None, None
 
     with tempfile.TemporaryDirectory() as output_dir:
         output_dir = Path(output_dir)
-        ic.write(output_dir, *dummy)
+        ic.write(output_dir, *DUMMY_ARGS)
 
         with open(output_dir / ic._file_name) as f:
             lines = f.readlines()
@@ -43,11 +44,10 @@ def test_initial_conditions_equilibrium_regrid(simple_2d_grid_with_subunits):
 
 def test_initial_conditions_percolation():
     ic = InitialConditionsPercolation()
-    dummy = None, None
 
     with tempfile.TemporaryDirectory() as output_dir:
         output_dir = Path(output_dir)
-        ic.write(output_dir, *dummy)
+        ic.write(output_dir, *DUMMY_ARGS)
 
         with open(output_dir / ic._file_name) as f:
             lines = f.readlines()
@@ -67,11 +67,10 @@ def test_initial_conditions_percolation_regrid(simple_2d_grid_with_subunits):
 
 def test_initial_conditions_rootzone_pressure_head():
     ic = InitialConditionsRootzonePressureHead(2.2)
-    dummy = None, None
 
     with tempfile.TemporaryDirectory() as output_dir:
         output_dir = Path(output_dir)
-        ic.write(output_dir, *dummy)
+        ic.write(output_dir, *DUMMY_ARGS)
 
         with open(output_dir / ic._file_name) as f:
             lines = f.readlines()
@@ -90,7 +89,6 @@ def test_initial_conditions_rootzone_regrid(simple_2d_grid_with_subunits):
 
 
 def test_initial_conditions_saved_state():
-    dummy = None, None
     with tempfile.TemporaryDirectory() as output_dir:
         output_dir = Path(output_dir)
         with open(output_dir / "foo.out", "w") as f:
@@ -98,19 +96,18 @@ def test_initial_conditions_saved_state():
 
         ic = InitialConditionsSavedState(output_dir / "foo.out")
 
-        ic.write(output_dir, *dummy)
+        ic.write(output_dir, *DUMMY_ARGS)
 
         assert Path(output_dir / "init_svat.inp").exists()
 
 
 def test_initial_conditions_saved_state_no_file():
-    dummy = None, None
     ic = InitialConditionsSavedState(r"doesntexist.txt")
 
     with tempfile.TemporaryDirectory() as output_dir:
         output_dir = Path(output_dir)
         with pytest.raises(FileNotFoundError):
-            ic.write(output_dir, *dummy)
+            ic.write(output_dir, *DUMMY_ARGS)
 
 
 def test_initial_conditions_saved_state_regrid():