+++ imod/tests/test_msw/test_landuse_options.py
@@ -55,7 +55,7 @@ def test_landuse_options(fixed_format_parser):
 
     with tempfile.TemporaryDirectory() as output_dir:
         output_dir = Path(output_dir)
-        lu_options.write(output_dir, None, None)
+        lu_options.write(output_dir, None, None, None, None)
 
         results = fixed_format_parser(
             output_dir / LanduseOptions._file_name, LanduseOptions._metadata_dict